requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiokafka==0.8.0
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorclass==2.2.2
colorlog==6.7.0
confluent-kafka==2.0.2
croniter==1.3.8
cryptography==39.0.0
Django==4.1.5
django-cron==0.6.0
django-encrypted-model-fields==0.6.5
djongo==1.3.6
dnspython==2.3.0
faust-streaming==0.10.4
frozenlist==1.3.3
funcy==1.18
graphql-core==3.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
intervaltree==3.1.0
jsonschema==4.17.3
kafka-python==2.0.2
mode==4.3.2
mode-streaming==0.3.5
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
opentracing==1.3.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-schema-registry-client==2.4.1
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4
strawberry-django-auth==0.374.0
strawberry-django-plus==1.35.2
strawberry-graphql==0.155.3
strawberry-graphql-django==0.9
terminaltables==3.1.10
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
venusian==3.0.0
yarl==1.8.2
